Apply changes on ridiplay/ing after forked from theothergrantdavidson/opencv-ts
- [clstudio] Upgrade biome to 1.4.1 (#487)

- Remove nginx from CLStudio web (#434)

- Adds support for detecting SFXs when creating toons from web.

- Revert opencv version.

  OpenCV 4.9 worked with unit tests under bun, but failed to work inside Photoshop
  plugin.

  The original version 4.5 seems to work.

  Maybe we need to modify the exporting code?

- Local changes to opencv-ts. (853e94f)

+++ src/core/Mat.d.ts
@@ -370,14 +370,14 @@ declare module Mat {
          * @param i0 A 0-based row index.
          * @returns the specified matrix row.
          */
-        intPtr(i0: number): Int8Array;
+        intPtr(i0: number): Int32Array;
         /**
          * Returns a pointer to the specified matrix row.
          * @param row Index along the dimension 0
          * @param col Index along the dimension 1
          * @returns the specified matrix row.
          */
-        intPtr(row: number, col: number): Int8Array;
+        intPtr(row: number, col: number): Int32Array;
         /**
          * Returns a pointer to the specified matrix row.
          * @param i0 A 0-based row index.
